// src/types/schemas/weaponlist.d.ts
/* prettier-ignore */
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Weapons = number[];

export interface Weaponlist {
	standard: Weapons;
	salmon: Weapons;
	grizzco: Weapons;
}